pageController decorators and new route extractor :D
