--- src/rest/base/apis.rest.ts
@@ -1,6 +1,0 @@
-import {Rest} from './rest';
-import {BlogRest} from '../blog.rest';
-
-export const ApisRest: Rest[] = [
-    ...BlogRest,
-];
